+++ sandbox/tester.f
@@ -1,4 +1,4 @@
 .( Start of tester.f )
-INCLUDE tester2.f .( back in tester on same line )
+INCLUDE ../sandbox/tester2.f .( back in tester on same line )
 foo
 .( end of tester.f )